Remove installation of special security provider
Do not replace the base Bouncy Castle security provider with the one in the app.

Use AES as the cipher for the sealed password objects instead of Blowfish.  The default provider does not include Blowfish, and it is unknown how to create the Cipher and all associated keys directly from a provider without registering the provider.