Fix crash when app is rotated after a TextInputLayout has cleared an error message
The error view for a TextInputLayout is not properly removed if setErrorEnabled(false) is called.  The view is removed, but after an orientation change, an exception occurs when restoring the view state.  Additionally, simply setting the error to null causes the error view to become invisible which maintains its location in the GUI.

Instead of calling setErrorEnabled(false), set the error field visibility manually.  An animation is used to queue after setErrorMsg's internal animation.