Added NFC Unlock functionality
The safe can now be unlocked using an NFC tag containing an
application/nfckey record that is an encrypted string.
The encrypted string is used as the passphrase to decrypt a value
stored in the application storage.
The decrypted value from application storage is the passphrase for the
encrypted string on the tag.
The decrypted tag value is the password for the pass safe.

(Specifically use application/nfckey mimetype to allow multiple records
so the same tag can be used for multiple nfckey purposed)